+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changes
 ~~~~~~~
 
-1.7.3 (unreleased)
+1.7.3 (2013-10-17)
 ------------------
 - launchpad #1232806: custom interpreter and openerp_scripts now work
   with OpenERP 6.1